+++ crates/git_hosting_providers/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "git"
+name = "git_hosting_providers"
 version = "0.1.0"
 edition = "2021"
 publish = false
@@ -9,33 +9,22 @@ license = "GPL-3.0-or-later"
 workspace = true
 
 [lib]
-path = "src/git.rs"
+path = "src/git_hosting_providers.rs"
 
 [dependencies]
 anyhow.workspace = true
 async-trait.workspace = true
-clock.workspace = true
-collections.workspace = true
-derive_more.workspace = true
-git2.workspace = true
+futures.workspace = true
+git.workspace = true
 gpui.workspace = true
-lazy_static.workspace = true
-log.workspace = true
-parking_lot.workspace = true
-rope.workspace = true
+isahc.workspace = true
+regex.workspace = true
 serde.workspace = true
-smol.workspace = true
-sum_tree.workspace = true
-text.workspace = true
-time.workspace = true
+serde_json.workspace = true
 url.workspace = true
 util.workspace = true
-windows.workspace = true
 
 [dev-dependencies]
 unindent.workspace = true
 serde_json.workspace = true
 pretty_assertions.workspace = true
-
-[features]
-test-support = []
