+++ crates/git_hosting_providers/src/providers/gitlab.rs
@@ -1,18 +1,16 @@
 use url::Url;
 
-use crate::{
-    BuildCommitPermalinkParams, BuildPermalinkParams, GitHostingProvider, ParsedGitRemote,
-};
+use git::{BuildCommitPermalinkParams, BuildPermalinkParams, GitHostingProvider, ParsedGitRemote};
 
-pub struct Gitee;
+pub struct Gitlab;
 
-impl GitHostingProvider for Gitee {
+impl GitHostingProvider for Gitlab {
     fn name(&self) -> String {
-        "Gitee".to_string()
+        "GitLab".to_string()
     }
 
     fn base_url(&self) -> Url {
-        Url::parse("https://gitee.com").unwrap()
+        Url::parse("https://gitlab.com").unwrap()
     }
 
     fn supports_avatars(&self) -> bool {
@@ -28,10 +26,10 @@ impl GitHostingProvider for Gitee {
     }
 
     fn parse_remote_url<'a>(&self, url: &'a str) -> Option<ParsedGitRemote<'a>> {
-        if url.starts_with("git@gitee.com:") || url.starts_with("https://gitee.com/") {
+        if url.starts_with("git@gitlab.com:") || url.starts_with("https://gitlab.com/") {
             let repo_with_owner = url
-                .trim_start_matches("git@gitee.com:")
-                .trim_start_matches("https://gitee.com/")
+                .trim_start_matches("git@gitlab.com:")
+                .trim_start_matches("https://gitlab.com/")
                 .trim_end_matches(".git");
 
             let (owner, repo) = repo_with_owner.split_once('/')?;
@@ -51,7 +49,7 @@ impl GitHostingProvider for Gitee {
         let ParsedGitRemote { owner, repo } = remote;
 
         self.base_url()
-            .join(&format!("{owner}/{repo}/commit/{sha}"))
+            .join(&format!("{owner}/{repo}/-/commit/{sha}"))
             .unwrap()
     }
 
@@ -65,7 +63,7 @@ impl GitHostingProvider for Gitee {
 
         let mut permalink = self
             .base_url()
-            .join(&format!("{owner}/{repo}/blob/{sha}/{path}"))
+            .join(&format!("{owner}/{repo}/-/blob/{sha}/{path}"))
             .unwrap();
         permalink.set_fragment(
             selection
@@ -81,116 +79,116 @@ mod tests {
     use super::*;
 
     #[test]
-    fn test_build_gitee_permalink_from_ssh_url() {
+    fn test_build_gitlab_permalink_from_ssh_url() {
         let remote = ParsedGitRemote {
-            owner: "libkitten",
+            owner: "zed-industries",
             repo: "zed",
         };
-        let permalink = Gitee.build_permalink(
+        let permalink = Gitlab.build_permalink(
             remote,
             BuildPermalinkParams {
-                sha: "e5fe811d7ad0fc26934edd76f891d20bdc3bb194",
+                sha: "e6ebe7974deb6bb6cc0e2595c8ec31f0c71084b7",
                 path: "crates/editor/src/git/permalink.rs",
                 selection: None,
             },
         );
 
-        let expected_url = "https://gitee.com/libkitten/zed/blob/e5fe811d7ad0fc26934edd76f891d20bdc3bb194/crates/editor/src/git/permalink.rs";
+        let expected_url = "https://gitlab.com/zed-industries/zed/-/blob/e6ebe7974deb6bb6cc0e2595c8ec31f0c71084b7/crates/editor/src/git/permalink.rs";
         assert_eq!(permalink.to_string(), expected_url.to_string())
     }
 
     #[test]
-    fn test_build_gitee_permalink_from_ssh_url_single_line_selection() {
+    fn test_build_gitlab_permalink_from_ssh_url_single_line_selection() {
         let remote = ParsedGitRemote {
-            owner: "libkitten",
+            owner: "zed-industries",
             repo: "zed",
         };
-        let permalink = Gitee.build_permalink(
+        let permalink = Gitlab.build_permalink(
             remote,
             BuildPermalinkParams {
-                sha: "e5fe811d7ad0fc26934edd76f891d20bdc3bb194",
+                sha: "e6ebe7974deb6bb6cc0e2595c8ec31f0c71084b7",
                 path: "crates/editor/src/git/permalink.rs",
                 selection: Some(6..6),
             },
         );
 
-        let expected_url = "https://gitee.com/libkitten/zed/blob/e5fe811d7ad0fc26934edd76f891d20bdc3bb194/crates/editor/src/git/permalink.rs#L7";
+        let expected_url = "https://gitlab.com/zed-industries/zed/-/blob/e6ebe7974deb6bb6cc0e2595c8ec31f0c71084b7/crates/editor/src/git/permalink.rs#L7";
         assert_eq!(permalink.to_string(), expected_url.to_string())
     }
 
     #[test]
-    fn test_build_gitee_permalink_from_ssh_url_multi_line_selection() {
+    fn test_build_gitlab_permalink_from_ssh_url_multi_line_selection() {
         let remote = ParsedGitRemote {
-            owner: "libkitten",
+            owner: "zed-industries",
             repo: "zed",
         };
-        let permalink = Gitee.build_permalink(
+        let permalink = Gitlab.build_permalink(
             remote,
             BuildPermalinkParams {
-                sha: "e5fe811d7ad0fc26934edd76f891d20bdc3bb194",
+                sha: "e6ebe7974deb6bb6cc0e2595c8ec31f0c71084b7",
                 path: "crates/editor/src/git/permalink.rs",
                 selection: Some(23..47),
             },
         );
 
-        let expected_url = "https://gitee.com/libkitten/zed/blob/e5fe811d7ad0fc26934edd76f891d20bdc3bb194/crates/editor/src/git/permalink.rs#L24-48";
+        let expected_url = "https://gitlab.com/zed-industries/zed/-/blob/e6ebe7974deb6bb6cc0e2595c8ec31f0c71084b7/crates/editor/src/git/permalink.rs#L24-48";
         assert_eq!(permalink.to_string(), expected_url.to_string())
     }
 
     #[test]
-    fn test_build_gitee_permalink_from_https_url() {
+    fn test_build_gitlab_permalink_from_https_url() {
         let remote = ParsedGitRemote {
-            owner: "libkitten",
+            owner: "zed-industries",
             repo: "zed",
         };
-        let permalink = Gitee.build_permalink(
+        let permalink = Gitlab.build_permalink(
             remote,
             BuildPermalinkParams {
-                sha: "e5fe811d7ad0fc26934edd76f891d20bdc3bb194",
+                sha: "b2efec9824c45fcc90c9a7eb107a50d1772a60aa",
                 path: "crates/zed/src/main.rs",
                 selection: None,
             },
         );
 
-        let expected_url = "https://gitee.com/libkitten/zed/blob/e5fe811d7ad0fc26934edd76f891d20bdc3bb194/crates/zed/src/main.rs";
+        let expected_url = "https://gitlab.com/zed-industries/zed/-/blob/b2efec9824c45fcc90c9a7eb107a50d1772a60aa/crates/zed/src/main.rs";
         assert_eq!(permalink.to_string(), expected_url.to_string())
     }
 
     #[test]
-    fn test_build_gitee_permalink_from_https_url_single_line_selection() {
+    fn test_build_gitlab_permalink_from_https_url_single_line_selection() {
         let remote = ParsedGitRemote {
-            owner: "libkitten",
+            owner: "zed-industries",
             repo: "zed",
         };
-        let permalink = Gitee.build_permalink(
+        let permalink = Gitlab.build_permalink(
             remote,
             BuildPermalinkParams {
-                sha: "e5fe811d7ad0fc26934edd76f891d20bdc3bb194",
+                sha: "b2efec9824c45fcc90c9a7eb107a50d1772a60aa",
                 path: "crates/zed/src/main.rs",
                 selection: Some(6..6),
             },
         );
 
-        let expected_url = "https://gitee.com/libkitten/zed/blob/e5fe811d7ad0fc26934edd76f891d20bdc3bb194/crates/zed/src/main.rs#L7";
+        let expected_url = "https://gitlab.com/zed-industries/zed/-/blob/b2efec9824c45fcc90c9a7eb107a50d1772a60aa/crates/zed/src/main.rs#L7";
         assert_eq!(permalink.to_string(), expected_url.to_string())
     }
 
     #[test]
-    fn test_build_gitee_permalink_from_https_url_multi_line_selection() {
+    fn test_build_gitlab_permalink_from_https_url_multi_line_selection() {
         let remote = ParsedGitRemote {
-            owner: "libkitten",
+            owner: "zed-industries",
             repo: "zed",
         };
-        let permalink = Gitee.build_permalink(
+        let permalink = Gitlab.build_permalink(
             remote,
             BuildPermalinkParams {
-                sha: "e5fe811d7ad0fc26934edd76f891d20bdc3bb194",
+                sha: "b2efec9824c45fcc90c9a7eb107a50d1772a60aa",
                 path: "crates/zed/src/main.rs",
                 selection: Some(23..47),
             },
         );
 
-        let expected_url = "https://gitee.com/libkitten/zed/blob/e5fe811d7ad0fc26934edd76f891d20bdc3bb194/crates/zed/src/main.rs#L24-48";
+        let expected_url = "https://gitlab.com/zed-industries/zed/-/blob/b2efec9824c45fcc90c9a7eb107a50d1772a60aa/crates/zed/src/main.rs#L24-48";
         assert_eq!(permalink.to_string(), expected_url.to_string())
     }
 }
