+++ crates/util/src/util.rs
@@ -1,7 +1,5 @@
 pub mod arc_cow;
-pub mod codeberg;
 pub mod fs;
-mod git_author;
 pub mod github;
 pub mod http;
 pub mod paths;
