+++ crates/zed/Cargo.toml
@@ -46,6 +46,8 @@ file_icons.workspace = true
 file_finder.workspace = true
 fs.workspace = true
 futures.workspace = true
+git.workspace = true
+git_hosting_providers.workspace = true
 go_to_line.workspace = true
 gpui.workspace = true
 headless.workspace = true
